fastapi/requirements.txt
anyio==3.5.0
asgiref==3.4.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.72.0
h11==0.12.0
httpcore==0.13.7
httptools==0.3.0
httpx==0.18.2
idna==3.3
motor==2.5.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pymongo==3.12.3
python-dotenv==0.19.2
python-jose==3.3.0
python-keycloak==0.25.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
